Se anade el editar una factura

+++ app/src/main/java/com/example/perraco/Objets/VentasObject.kt
@@ -5,6 +5,6 @@ import java.io.Serializable
 data class VentasObjeto(
     var _id:Int,
     var fecha:String,
-    var totalVenta:Int,
+    var totalVenta:Double,
     var articulos:List<ArticuloObjeto>
 ): Serializable
